robot-mall-server/requirements.txt
docutils==0.14
pillow==5.4.1
pytesseract==0.2.6
requests==2.21.0
robot-jsonschemalibrary==0.1.1
robotframework==3.1.1
urllib3==1.24.1
jsonschema==2.6.0
xlwt==1.3.0
xlrd==1.2.0